Stage 2 target achieved.
Now we replace the block-fading assumption with the inter-symbol independent fading assumption so we can get a more smooth waterfall curve with the same amount of monte-carlo simulation.

Ready to draw the water-fall curve. Also taking into account of the noise-power estimation error (in this way we can save some simulation actually :) ).